Move ErrorComponents to a separate directory
