+++ frontend/src/components/ViewActivities/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import app from '../Firebase';
 import { getUserUid } from '../AuthUtils';
 import ActivityList from './activitylist.js';
-import * as ErrorComponents from './ErrorComponents.js';
+import * as ErrorComponents from '../Errors';
 
 import * as DB from '../../constants/database.js';
